+++ apps/backend/src/modules/auth/auth.controller.ts
@@ -1,3 +1,5 @@
+import { type MultipartFile } from "@fastify/multipart";
+
 import { APIPath } from "~/libs/enums/enums.js";
 import {
 	type APIHandlerOptions,
@@ -6,6 +8,7 @@ import {
 } from "~/libs/modules/controller/controller.js";
 import { HTTPCode } from "~/libs/modules/http/http.js";
 import { type Logger } from "~/libs/modules/logger/logger.js";
+import { type FileUploadResponseDto } from "~/modules/files/libs/types/types.js";
 import {
 	type AuthenticatedUserPatchRequestDto,
 	authenticatedUserPatchValidationSchema,
@@ -279,6 +282,12 @@ class AuthController extends BaseController {
 		this.authService = authService;
 
 		this.addRoute({
+			handler: this.uploadAvatar.bind(this),
+			method: "POST",
+			path: AuthApiPath.AUTHENTICATED_USER_UPLOAD_AVATAR,
+		});
+
+		this.addRoute({
 			handler: this.signUp.bind(this),
 			method: "POST",
 			path: AuthApiPath.SIGN_UP,
@@ -473,6 +482,63 @@ class AuthController extends BaseController {
 		const { body } = options;
 
 		const data = await this.authService.signUp(body);
+
+		return {
+			payload: { data },
+			status: HTTPCode.CREATED,
+		};
+	}
+
+	/**
+	 * @swagger
+	 * /auth/authenticated-user/avatar:
+	 *   post:
+	 *     summary: Upload an avatar
+	 *     description: Upload an avatar to the authorized user
+	 *     tags:
+	 *       - Auth
+	 *     security:
+	 *       - bearerAuth: []
+	 *     requestBody:
+	 *       required: true
+	 *       content:
+	 *         multipart/form-data:
+	 *           schema:
+	 *             type: object
+	 *             required:
+	 *               - file
+	 *             properties:
+	 *               file:
+	 *                 type: string
+	 *                 format: binary
+	 *                 description: File to upload (image/jpeg or image/png)
+	 *     responses:
+	 *       201:
+	 *         description: Avatar updated succesfully
+	 *         content:
+	 *           application/json:
+	 *             schema:
+	 *               type: object
+	 *               properties:
+	 *                 data:
+	 *                   $ref: '#/components/schemas/FileUploadResponseDto'
+	 */
+
+	public async uploadAvatar(
+		options: APIHandlerOptions<{
+			body: {
+				file: MultipartFile;
+			};
+			user: UserAuthResponseDto;
+		}>,
+	): Promise<APIHandlerResponse<FileUploadResponseDto>> {
+		const { body, user } = options;
+
+		if (!user) {
+			throw new AuthError();
+		}
+
+		const data = await this.authService.uploadAvatar(user.id, body.file);
 
 		return {
 			payload: { data },