+++ apps/frontend/src/libs/components/icon/libs/maps/icon-name-to-svg.map.ts
@@ -9,6 +9,7 @@ import Error from "~/assets/images/icons/error.svg?react";
 import EyeOff from "~/assets/images/icons/eye-off.svg?react";
 import Eye from "~/assets/images/icons/eye.svg?react";
 import Link from "~/assets/images/icons/link.svg?react";
+import Logout from "~/assets/images/icons/logout.svg?react";
 import Map from "~/assets/images/icons/map.svg?react";
 import Message from "~/assets/images/icons/message.svg?react";
 import Route from "~/assets/images/icons/route.svg?react";
@@ -32,6 +33,7 @@ const iconNameToSvg: Record<
 	eye: Eye,
 	eyeOff: EyeOff,
 	link: Link,
+	logout: Logout,
 	map: Map,
 	message: Message,
 	route: Route,