+++ apps/frontend/src/libs/types/types.ts
@@ -1,4 +1,5 @@
 export { type AsyncThunkConfig } from "./async-thunk-config.type.js";
+export { type DropdownItem } from "./dropdown-item.type.js";
 export { type IconName } from "./icon-name.type.js";
 export { type Location } from "./location.type.js";
 export { type NavigationItem } from "./navigation-item.type.js";
@@ -9,6 +10,7 @@ export {
 	type APIResponse,
 	type APIValidationErrorDetail,
 	type Coordinates,
+	type FileUploadResponseDto,
 	type PaginationMeta,
 	type PointGeometry,
 	type ValidationSchema,