+++ apps/frontend/src/modules/files/libs/enums/enums.ts
@@ -1,1 +1,0 @@
-export { FileFolderName, FilesApiPath } from "@smartscapes/shared";