+++ apps/frontend/src/modules/files/libs/types/types.ts
@@ -1,4 +1,0 @@
-export {
-	type FileUploadRequestDto,
-	type FileUploadResponseDto,
-} from "@smartscapes/shared";