+++ apps/frontend/src/modules/routes/libs/types/types.ts
@@ -1,6 +1,8 @@
+export { type PatchActionPayload } from "./patch-action-payload.type.js";
 export {
 	type RouteConstructRequestDto,
 	type RouteConstructResponseDto,
 	type RouteFindAllOptions,
 	type RouteGetByIdResponseDto,
+	type RoutePatchRequestDto,
 } from "@smartscapes/shared";