+++ apps/frontend/src/modules/routes/routes-api.ts
@@ -1,10 +1,4 @@
-import {
-	ContentType,
-	type RouteFindAllOptions,
-	type RouteGetByIdResponseDto,
-} from "@smartscapes/shared";
-
-import { APIPath } from "~/libs/enums/enums.js";
+import { APIPath, ContentType } from "~/libs/enums/enums.js";
 import { BaseHTTPApi } from "~/libs/modules/api/api.js";
 import { type HTTP } from "~/libs/modules/http/http.js";
 import { type Storage } from "~/libs/modules/storage/storage.js";
@@ -12,8 +6,11 @@ import { type APIResponse } from "~/libs/types/types.js";
 
 import { RoutesApiPath } from "./libs/enums/enums.js";
 import {
+	type PatchActionPayload,
 	type RouteConstructRequestDto,
 	type RouteConstructResponseDto,
+	type RouteFindAllOptions,
+	type RouteGetByIdResponseDto,
 } from "./libs/types/types.js";
 
 type Constructor = {
@@ -67,6 +64,23 @@ class RoutesApi extends BaseHTTPApi {
 			{
 				hasAuth: false,
 				method: "GET",
+			},
+		);
+
+		return await response.json();
+	}
+
+	public async patchRoute({
+		id,
+		payload,
+	}: PatchActionPayload): Promise<APIResponse<RouteGetByIdResponseDto>> {
+		const response = await this.load<APIResponse<RouteGetByIdResponseDto>>(
+			this.getFullEndpoint(RoutesApiPath.$ID, { id: id.toString() }),
+			{
+				contentType: ContentType.JSON,
+				hasAuth: true,
+				method: "PATCH",
+				payload: JSON.stringify(payload),
 			},
 		);