+++ apps/frontend/src/modules/routes/routes.ts
@@ -11,7 +11,10 @@ const routeApi = new RoutesApi({
 });
 
 export { routeApi };
-export { type RouteGetByIdResponseDto } from "./libs/types/types.js";
+export {
+	type RouteGetByIdResponseDto,
+	type RoutePatchRequestDto,
+} from "./libs/types/types.js";
 export {
 	actions,
 	constructRouteReducer,