+++ packages/shared/src/modules/files/libs/enums/files-api-path.enum.ts
@@ -1,6 +1,6 @@
 const FilesApiPath = {
 	ROOT: "/",
-	UPLOAD: "/upload/:folder",
+	UPLOAD: "/upload/:folder/:entityId",
 } as const;
 
 export { FilesApiPath };